PluginManager finished
