+++ app/src/main/java/tickets/file_dao/FileAccess.java
@@ -1,14 +1,6 @@
-package tickets.server.dataAccess.FileDAO;
+package tickets.file_dao;
 
-        import java.io.File;
         import java.io.IOException;
-        import java.nio.charset.Charset;
-        import java.nio.file.Files;
-        import java.nio.file.Path;
-        import java.nio.file.Paths;
-        import java.nio.file.StandardOpenOption;
-        import java.util.ArrayList;
-        import java.util.Arrays;
         import java.util.List;
 
 public class FileAccess {
