+++ app/src/main/java/tickets/file_dao/UserDA_File.java
@@ -1,9 +1,9 @@
-package tickets.server.dataAccess.FileDAO;
+package tickets.file_dao;
 
 import java.util.List;
 
-import tickets.server.dataAccess.DataAccess;
-import tickets.server.dataAccess.Interfaces.UserDataAccess;
+import tickets.server.dataaccess.DataAccess;
+import tickets.server.dataaccess.interfaces.UserDataAccess;
 
 /**
  * Created by Pultilian on 4/12/2018.
