+++ app/src/main/java/tickets/relational_dao/UserDA_Rel.java
@@ -1,10 +1,10 @@
-package tickets.server.dataAccess.RelationalDAO;
+package tickets.relational_dao;
 
 import java.sql.ResultSet;
 import java.util.List;
 
-import tickets.server.dataAccess.DataAccess;
-import tickets.server.dataAccess.Interfaces.UserDataAccess;
+import tickets.server.dataaccess.DataAccess;
+import tickets.server.dataaccess.interfaces.UserDataAccess;
 
 /**
  * Created by Pultilian on 4/12/2018.
