+++ app/src/main/java/tickets/server/dataaccess/DataAccess.java
@@ -1,10 +1,9 @@
-package tickets.server.dataAccess;
+package tickets.server.dataaccess;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
-import java.util.List;
 
 /**
  * Created by Pultilian on 4/9/2018.
